seperated script to bash  bash command
